# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AdminUserDTO"
                  },
                  {
                    "$ref": "#/components/schemas/StaffUserDTO"
                  },
                  {
                    "$ref": "#/components/schemas/CustomerUserDTO"
                  }
                ],
                "discriminator": {
                  "propertyName": "role",
                  "mapping": {
                    "admin": "#/components/schemas/AdminUserDTO",
                    "staff": "#/components/schemas/StaffUserDTO",
                    "customer": "#/components/schemas/CustomerUserDTO"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/stores": {
      "post": {
        "operationId": "StoresController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Loja criada com sucesso"
          }
        },
        "summary": "Criar loja",
        "tags": [
          "Stores"
        ]
      }
    },
    "/stores/{id}": {
      "put": {
        "operationId": "StoresController_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDTO"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Loja atualizada com sucesso"
          }
        },
        "summary": "Atualizar loja",
        "tags": [
          "Stores"
        ]
      }
    }
  },
  "info": {
    "title": "Raízes do Nordeste API",
    "description": "API usando NestJS, Prisma",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AdminUserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Nome completo"
          },
          "email": {
            "type": "string",
            "example": "johndoe@example.com",
            "description": "E-mail"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Senha"
          },
          "role": {
            "type": "string",
            "example": "admin",
            "enum": [
              "admin",
              "staff",
              "customer"
            ],
            "description": "Role do usuário"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "StaffUserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Nome completo"
          },
          "email": {
            "type": "string",
            "example": "johndoe@example.com",
            "description": "E-mail"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Senha"
          },
          "role": {
            "type": "string",
            "example": "staff",
            "enum": [
              "admin",
              "staff",
              "customer"
            ],
            "description": "Papel do usuário"
          },
          "profile": {
            "type": "string",
            "enum": [
              "kitchen",
              "waiter",
              "cashier"
            ],
            "description": "Perfil (obrigatório para staff)"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role",
          "profile"
        ]
      },
      "CustomerDataDTO": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string",
            "example": "123.456.789-00",
            "description": "CPF do cliente"
          },
          "consent": {
            "type": "boolean",
            "example": true,
            "description": "Consentimento do cliente"
          },
          "consentAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00Z",
            "description": "Data do consentimento"
          },
          "points": {
            "type": "number",
            "example": 0,
            "description": "Pontos acumulados"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00Z",
            "description": "Data de criação"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00Z",
            "description": "Data de atualização"
          }
        },
        "required": [
          "cpf",
          "consent",
          "consentAt",
          "points",
          "createdAt",
            "updatedAt"
        ]
      },
      "CustomerUserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Nome completo"
          },
          "email": {
            "type": "string",
            "example": "johndoe@example.com",
            "description": "E-mail"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Senha"
          },
          "role": {
            "type": "string",
            "example": "customer",
            "enum": [
              "admin",
              "staff",
              "customer"
            ],
            "description": "Papel do usuário"
          },
          "customerData": {
            "description": "Dados do cliente",
            "allOf": [
              {
                "$ref": "#/components/schemas/CustomerDataDTO"
              }
            ]
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role",
          "customerData"
        ]
      },
      "AddressDTO": {
        "type": "object",
        "properties": {
          "streetAndNumber": {
            "type": "string",
            "example": "Rua das Flores, 123",
            "description": "Rua e número"
          },
          "cep": {
            "type": "string",
            "example": "12345-678",
            "description": "CEP"
          },
          "neighborhood": {
            "type": "string",
            "example": "Bairro Central",
            "description": "Bairro"
          },
          "city": {
            "type": "string",
            "example": "Fortaleza",
            "description": "Cidade"
          }
        },
        "required": ["streetAndNumber", "cep", "neighborhood", "city"]
      },
      "CreateStoreDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Loja do Nordeste",
            "description": "Nome da loja"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDTO",
            "description": "Endereço dividido em campos"
          }
        },
        "required": ["name", "address"]
      },
      "UpdateStoreDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Loja Atualizada",
            "description": "Novo nome da loja"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDTO",
            "description": "Novo endereço dividido em campos"
          }
        }
      }
    }
  }
}